requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appscript==1.2.2
click==8.1.7
cobble==0.1.3
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.3
lxml==4.9.3
MarkupSafe==2.1.3
pdfkit==1.0.0
pydantic_core==2.10.1
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wkhtmltopdf==0.2
